Use bundled JCE policy and remove alternatives entry for JCE policy
Prior to the 152 release of Java 8 only the limited JCE policy
was shipped with the JDK and the unlimited JCE policy was shipped
separately and had to be installed manually. The RPM had taken
care of this, and also set up the policy in use using the
alternatives system. The chosen policy was applied to all running
apps.

With the 152 release both the limited and unlimited policies ship
with the JDK and it is possible to specify which is used on a per
app basis, as well as set a system default. This removes the need
to set up the policy to use the alternatives system.

See:
https://golb.hplar.ch/p/JCE-policy-changes-in-Java-SE-8u151-and-8u152